Configure Scaffolder (#4)
### 📑 Summary

This pull request introduces a set of changes to improve the development
workflow and add new scaffolding capabilities to the monorepo. The most
important changes include updates to the `README.md` file, additions to
the `package.json` file, and the creation of several utility functions
and configurations for scaffolding.

### Documentation and Configuration Updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R22-R45):
Added sections for development, including code quality checks (linting
and auto-formatting) and scaffolding instructions.
*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L10-R19):
Added the `generate` script and included `@turbo/gen` as a dev
dependency.

### Scaffolding Enhancements:

*
[`turbo/generators/config.ts`](diffhunk://#diff-95d7e92db69fdcaffd10a98262184cb49a242e9d83ae4ef9e99ec704a16573f9R1-R6):
Created a configuration file for the generator, setting up a custom
action type.
*
[`turbo/generators/types.ts`](diffhunk://#diff-b452c79c710883cc506e8159dff94b78a454df89426c63ebcb9d061c34355128R1-R7):
Defined TypeScript types for the custom action configuration used in the
generator.

### Utility Functions for Scaffolding:

*
[`turbo/generators/utilities/cwd.ts`](diffhunk://#diff-bfc012aa34fa90022db24f29c92c01947ff7af9a658fc04c8815f68c97a00887R1-R5):
Added a utility function to get the current working directory path.
*
[`turbo/generators/utilities/getAllDirectories.ts`](diffhunk://#diff-7943009a3983ef997a8c139264a5f9380c2ca46c35dd817f2a420edb529500fdR1-R14):
Implemented a function to retrieve all directories within a given path.
*
[`turbo/generators/utilities/getAllFilesRecursivelySync.ts`](diffhunk://#diff-fcd9f9e548157c2bff38c7c4b0a08210045e4ef54da1105b0c3fd8e5451399d1R1-R20):
Created a function to get all files recursively within a directory.
*
[`turbo/generators/utilities/getTemplateFiles.ts`](diffhunk://#diff-ec8814bdfe5e5ff131a73deeee82e56b7d4e9325b5fdfee0c55d0ed45a6be3e4R1-R8):
Added a function to get template files from a base path.
*
[`turbo/generators/utilities/normalizeTemplateName.ts`](diffhunk://#diff-6574edbf33f126d2f426e62a622a48b04eda3fa2bc76b309e95799a994f21f0cR1-R3):
Implemented a function to normalize template names by replacing specific
patterns.
*
[`turbo/generators/utilities/root.ts`](diffhunk://#diff-669b01661c9897c4d975b46c0c83e76ed16765b94af8a88aefe2bdc2c2ce6b41R1-R5):
Added a utility function to get the root directory path.
*
[`turbo/generators/utilities/runActionType.ts`](diffhunk://#diff-0454809719c392639e85ecd1be2cd9f47f6a81bf79b5d5e73579e5364391d486R1-R12):
Defined a custom action type for running callback functions in the
generator.
*
[`turbo/generators/utilities/templates.ts`](diffhunk://#diff-40e787932c3ac689129b37e4c3267e2fd1cb8cdceefb1aec2db803980e99748aR1-R5):
Created a utility function to get paths to template files.

### 🎩 Testing instructions

- Running `pnpm generate` should trigger the scaffolder

### ✅ Checks

- [x] My pull request adheres to the code style of this project
- [x] My code requires changes to the documentation
- [x] I have updated the documentation as required
- [x] All the tests have passed