Setup Changesets (#6)
### 📑 Summary

This pull request includes several changes to the project configuration
and the creation of a new core package. The most important changes are
summarized below:

### Configuration Updates:
* Added a new configuration file `.changeset/config.json` to manage
versioning and changelogs using the Changesets tool.
* Updated `package.json` to include a new script for version bumping
(`version:bump`) and added `@changesets/cli` to the `devDependencies`.

### Core Package Creation:
* Created a new core package with the name `@micra/core` and documented
the creation in a changeset file `.changeset/pink-queens-confess.md`.
* Updated the version of the new core package in
`packages/core/package.json` to `0.0.0`.

### 🎩 Testing instructions

- Running `pnpm version:bump` should guide the developer to generate a
new changeset

### ✅ Checks

- [x] My pull request adheres to the code style of this project
- [x] My code requires changes to the documentation
- [x] I have updated the documentation as required
- [x] All the tests have passed

### ℹ Additional Information

This does not configure the continuous delivery and GitHub action for
Changesets. For now we won't be publishing the packages and we can
tackle this in at a later time.