Tweaks and fixes to docs and package.json (#7)
### 📑 Summary

This pull request includes several updates to the documentation and
metadata files to improve the project's accessibility and
maintainability. The most important changes include updating URLs in the
`README.md` files to point to the latest version on GitHub, adding
author and repository information to the `package.json` file, and
removing unnecessary whitespace.

Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L2-R8):
Updated image and link URLs to point to the latest version on GitHub.
*
[`packages/core/README.md`](diffhunk://#diff-af0c08d6f291aa75c55d4ad3ec19fd4b728edcf43a39c80d566f661b3b311c44L6-R8):
Updated documentation, contributing, and code of conduct links to point
to the latest version on GitHub.

Metadata improvements:

*
[`packages/core/package.json`](diffhunk://#diff-0b810c38f3c138a3d5e44854edefd5eb966617ca84e62f06511f60acc40546c7R6-R13):
Added author, repository, and bugs information to the package metadata.

Whitespace cleanup:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L62):
Removed unnecessary whitespace in the documentation section.

### ✅ Checks

- [x] My pull request adheres to the code style of this project
- [x] My code requires changes to the documentation
- [x] I have updated the documentation as required
- [x] All the tests have passed